OH_K12_clean
Hey guys, I have cleaned the data and exported to a new excel sheet for later use. The R code of how I cleaned it is also included in the R file. Please feel free to refer to it!! Currently the missing data is regarding dates (which I haven't figured out how to fix, or maybe we don't fix it), and also enrollment. I need to wait for Annika's response for the documentary because I feel like enrollment and enrollment_1 are related thus I can fix the missing values accordingly. Just to take some notes here!